# Add Betsy to speed up BC6 compression