--- a/modules/betsy/config.py
+++ b/modules/betsy/config.py
@@ -1,0 +1,6 @@
+def can_build(env, platform):
+    return env.editor_build
+
+
+def configure(env):
+    pass