--- a/servers/rendering_server.cpp
+++ b/servers/rendering_server.cpp
@@ -3518,6 +3518,7 @@ void RenderingServer::init() {
 	// See `const bool can_s3tc_bptc` in the resource importer.
 	GLOBAL_DEF_RST("rendering/textures/vram_compression/import_s3tc_bptc", false);
 	GLOBAL_DEF_RST("rendering/textures/vram_compression/import_etc2_astc", false);
+	GLOBAL_DEF("rendering/textures/vram_compression/compress_with_gpu", true);
 
 	GLOBAL_DEF("rendering/textures/lossless_compression/force_png", false);