… doc/classes/ProjectSettings.xml
… modules/betsy/bc6h.glsl
… thirdparty/README.md
… thirdparty/betsy/LICENSE.md