Excerpt of #91150.
Fixes #80456.

Implements the <a href=https://github.com/darksylinc/betsy>Betsy GPU texture compressor</a> to drastically speed up BC6 compression, which will be essential for compressing baked lightmaps.
 
Performance (compression ONLY):
Build - Production 64 bit,
CPU - Ryzen 9 5900x,
GPU - RTX 4060 TI
| image | CVTT | Betsy |
|--------|------|-------|
| <a href=https://polyhaven.com/a/symmetrical_garden_02>Symmetrical garden 8k .hdr with mipmaps</a> | 92.4s  | 475ms |
| <a href=https://polyhaven.com/a/cobblestone_street_night>Cobblestone Street Night .hdr 4k with mipmaps</a>| 26.5s | 217ms |
| <a href=https://polyhaven.com/a/laufenurg_church>Laufenurg Church 8k .hdr with mipmaps</a>| 99.3s | 440ms |
| <a href=https://polyhaven.com/a/little_paris_eiffel_tower>Little Paris 8k .hdr with mipmaps</a>| 92.7s | 467ms |

TODO:
- [x] optimize the setup (shader compilation, RGB->RGBA conversion), since that takes up most of the time,

TODO after merging:
- move the optimized functions to `Image` (#92291 and #92799),
- Optimize .hdr file loading (#95291),
- implement signed compression support (or at least try to),
- fix layered texture import when compressing layers with different signs,
- look into optimizing CVTT to be a viable fallback.
- parallelize mipmap compression (idea: pack all mips into a second image with half the width and the same height as the original and run in parallel).