@BlueCube3310 2024-05-04
Signed compression is now supported, though because the sign detection happens on individual layers, most layered images break when combined.