@fire 2024-05-04
Is there a way of bc6 using sign detection on individual layers or the entire set at once without too much research time?

This seems like we need to implement a batch system?