@BlueCube3310 2024-05-06
Did some benchmarks, the difference in performance is incredible (up to 155 times faster on some large images).