@BlueCube3310 2024-05-06
> It's strange how my import speeds are quite a bit slower than @BlueCube3310's though, considering my GPU is significantly faster.

My results only contain the time it takes to execute the `compress` function, and not the whole import process (decoding, mipmap generation, etc.). On my machine the whole process took ~14s.