@fire 2024-05-07
I was thinking about `fix signed compression (when importing a layered texture, some slices are signed and others aren't, so the combined texture breaks).` can we use disable the signed path or somehow resolve or do we have to implement batched signing properly?