@BlueCube3310 2024-05-07
> I was thinking about `fix signed compression (when importing a layered texture, some slices are signed and others aren't, so the combined texture breaks).` can we disable the signed path or somehow resolve or do we have to implement batched signing properly?

As far as I'm aware, directional lightmaps depend on signed float data for storing direction information, so forcing the texture to be compressed as unsigned will cause artifacts to appear.

I also found out today that Betsy doesn't really support the signed variant of bc6, so it will need to be properly implemented first.