@fire 2024-05-07
Adding signed compression to betsy doesn't sound trivial.