@BlueCube3310 2024-05-08
> Adding signed compression to betsy doesn't sound trivial.

I'll see if I can get it to work by July. If not, signed images will either be uncompressed or converted by CVTT as a fallback.