@BlueCube3310 2024-05-09
The GPU compressor can now be toggled from the project settings and the compressor works in rendering modes other than Forward+.