@BlueCube3310 2024-05-09
So I tried compressing a directional lightmap with Betsy and noticed some weird artifacts:
![betsy](https://github.com/godotengine/godot/assets/53150244/c4475b9c-6ffb-4bc0-afc0-a859e01c9d22)

I thought this was a compressor-specific issue, so I tried CVTT (the one used by Godot):
![cvtt](https://github.com/godotengine/godot/assets/53150244/2e1d7681-bf2e-4531-af53-589fcc388f1a)

I have no idea why this happens. It could be a format limitation, or simply a shared issue between the compressors. Either way, directional lightmaps should probably remain uncompressed.