@jcostello 2024-05-09
Ohh, that is a shame. Directional Lightmaps are so heavy. Maybe for moment you can use compression on non-directional lightmaps until de issue is solved