@fire 2024-05-09
Can you compress some thing else that has signed floats for comparison and against some official signed float bc6 encoder?