@BlueCube3310 2024-05-09
> Can you compress some thing else that has signed floats for comparison and against some stable signed float bc6 encoder?

![bc6](https://github.com/godotengine/godot/assets/53150244/19b5586d-bf37-4dc3-b7f1-432dff9039b2)

Left: Microsoft's DirectXTex,
Middle: CVTT,
Right: Betsy