@fire 2024-05-09
Is there a quality knob we can maximize for Betsy before we try too hard on debugging?