@akien-mga 2024-05-10
CC @darksylinc, FYI. Finally putting Betsy to good use in Godot!