@BlueCube3310 2024-05-12
It looks like ASTC doesn't handle negative values either.