@fire 2024-05-13
Been playing around with a python script and chatgpt4.

```py
import numpy as np

def ulp(x):
    return np.nextafter(x, np.inf) - x

def map_minus_one_to_one(val):
    return val * 0.5 + 0.5 # Mapping [-1; 1] -> [0; 1]

def map_one_to_zero(val):
    return 1.0 - val * 0.5 + 0.5; # Mapping [-1; 1] -> [1; 0]

def map_exp(val):
    return np.exp(val)

def map_exp2(val):
    return np.exp2(val)

def map_exp_inverse(val):
    return np.exp(1 - val)

values = np.logspace(-3, 3, 1000)
mappings = {'map_minus_one_to_one': map_minus_one_to_one, 
            'map_one_to_zero': map_one_to_zero, 
            'map_exp': map_exp, 
            'map_exp2': map_exp, 
            'map_exp_inverse': map_exp_inverse}

for name, mapping in mappings.items():
    mapped_values = mapping(values)
    ulps = ulp(mapped_values)
    print(f"Mapping {name}:")
    print(f"Min ULP: {np.min(ulps)}")
    print(f"Max ULP: {np.max(ulps)}")
    print(f"Mean ULP: {np.mean(ulps)}")
    print()
```

Mapping map_minus_one_to_one:
Min ULP: 1.1102230246251565e-16
Max ULP: 5.684341886080802e-14
Mean ULP: 5.71298564011613e-15

Mapping map_one_to_zero:
Min ULP: 2.168404344971009e-19
Max ULP: 5.684341886080802e-14
Mean ULP: 5.614703363701623e-15

Mapping map_exp:
Min ULP: nan
Max ULP: nan
Mean ULP: nan

Mapping map_exp2:
Min ULP: nan
Max ULP: nan
Mean ULP: nan

Mapping map_exp_inverse:
Min ULP: 5e-324
Max ULP: 4.440892098500626e-16
Mean ULP: 2.0963494167660428e-16