@darksylinc 2024-05-13
That's not exactly what I meant (although it's not bad, and it's close in spirit).

What I meant is that the average use case is going to use certain ranges more than others; thus the testing would need to generate several lightmaps and compare using real world data.