@lyuma 2024-05-13
My understanding is BC6H and ASTC hdr allow negative values.
I don't think we should change the format of dir lightmaps. We should debug the betsy shader code.

We should double check it isn't a shader optimizer issue. For example, in HLSL (some versions of)? fxc / dx11 will assume that reads from `Texture2D<float4>` return positive values. Something like `float4 x = tex2D(...); if (x.x <0) { ... }` will be optimized out or `float foo = max(0.0, tex2D(...).x)` will be optimized to `float foo = tex2D(...)`
Maybe GLSL optimizers have a similar issue