@clayjohn 2024-05-14
Once this is added, we should look into compressing ReflectionProbes and Skys when the update mode is set to "update once"